+++ tests/test_extractor.py
@@ -1,29 +1,14 @@
-import boto3
 import botocore
 import json
 import os
 from botocore.client import BaseClient
-from mock import MagicMock
 from moto import mock_s3
 from pytest import fixture
 from typing import Callable, List, Tuple, Union
 
-from email_parser import ParsedEmail
+from extractor.email_parser import ParsedEmail
 from extractor.extractor import lambda_handler
-from.utils import create_s3_object
-
-
-@fixture
-def mock_sns_client(mocker) -> botocore.client.BaseClient:
-    real_client_factory = boto3.client
-    mock_client = MagicMock()
-
-    def factory_factory(client_type: str):
-        return mock_client if client_type == 'sns' else real_client_factory(client_type)
-
-    mocker.patch.object(boto3, 'client', side_effect=factory_factory)
-
-    return mock_client
+from .utils import create_s3_object
 
 
 @fixture
@@ -71,5 +56,5 @@ def test_that_lambda_handler_wires_everything_correctly(
     lambda_handler(event, None)
     mock_sns_client.publish.assert_called_once_with(
         TargetArn=sns_arn,
-        Message=json.dumps({'default': json.dumps(ParsedEmail.from_bytes(base_email).as_dict())}),
+        Message=json.dumps({'default': ParsedEmail.from_bytes(base_email).as_json()}),
         MessageStructure='json')
